Fix some bugs in adding+deleting characters to text display
Fix bug with the offsets of the letters of the text selection.  Insertion of letters in middle of sentence is still bugged, but will be resolved by taking the  _numTChar array and creating a binary index tree to get the sum of a range of values from this array. This binary index tree will be used in the addWord and deleteWord functions to gget the appropriate insertion point for new characters.