Refactor code for TextDisplay
Refactored 2 things:I  stopped the prop drilling from mainBody to TextDisplay by removing the function callbacks and adding the function calls of AppBrain directly to the TextDisplay widget.   I also refactored the CopyAll button by creating a TextDisplayButton class.  This class will be used for the Paste and Undo buttons.