Improve drawing pad so curves are smoother
Created a class, BezierInterpolationND that is used to take the position points in every stroke and interpolate them with cubic splines.  The Writing pad takes the functions and draws them.  The interpolation function is a bit unstable, and required a quick fix.  Each stroke is split up into smaller strokes that are interpolated.