Refactor main adn main_body to make InfoScreen popup
Put some widgets from Main to MainBody and stopped using Multiprovider in place of a regular ChangeNotifierProvider.  I Put the instantiation of AppBrain higher up the widget tree to account for the state of the InfoScreen:  The main app now has is a stack wit h the main chunk of the app, as well as an information screen that will be transparent and go over the app.  The actual rendering of the app will be toggled by the info button at the top right of the screen.