Fix info screen text being in wrong place on Android
Refactored the InfoScreen by changing how it takes data from info_screen_content.  The second value of an element of the list infoScreenPageContents is now in units of the length between the bottom of the screen and the bottom of the Text screen in TextDisplay.  The y value does not correspond to the original screen coordinates with a downward facing y axis. It is now reversed.  This is apparent in the InfoScreen build function.