Fix InfoScreen bug where text would be in wrong position
The previous changes changed how the data from info_screen_content is read for the y position of the aligned texts.  This data is now read as the height of the text relative to the bottom of the screen.  Its units is in bottomWidgetsHeightRatio.  I fixed a bug in infoScreen.dart where the bottomWidgetsHeightRatio was incorrectly calculated, messing up the positioning of the texts in infoScreen.  This would also mess up the positions in other devices of different dimensions.  The error was that the Alignment() function makes sure the aligned widget is always inside its parent, whereas I thought that the widget could be completely out of its parent if you were not careful enough.