Create code that animates simple strokes for Practice Mode
I made a way to animate each stroke.  The animations need to be adjusted, but the main strokes can be created.  Loop-de-loops still need to be accounted for.  The PracticeCharacterPage and DrawnStroke widgets are simplified.  The GradientData class in step_by_step_stroke_data has all the animation data, and makes the shadercallback function that is used in the ShaderMask widget in DrawnStroke.  The GradientData class stores the animation info for each stroke in a character.  The characterToGradientData map uses the GradientData class to store the relevant data on how to animate the stroke.  The PracticeCharacterPage uses one entry of characterGradientData and takes gives the string character of the current stroke as well as the shadercallback to the DrawnStroke widget