Fix errors in sweep gradient animations
The sweep animations were not working for multiple reasons.  A syntactical bug in GradientData kept passing the GradientRotationReverse transform for every sweep animation, and not when isReversed = true.  The semantic error of assuming the coordinate system for the GradientTransform was located in the center of the targeted widget as well as forgetting the coordinates had a Y-axis pointing downwards to the bottom of the screen led to a gradient tranform that completely wrong, usually sending its center offscreen.  \n Also, because the y-axis is flipped, the sweeping angle was instead clockwise instead of counterclockwise.  This must me accounted for in the characterToGradientData map in step_by_step_stroke_data