Fix sweep gradient animations
The sweep gradient cannot handle negative angles or angles greater than 2pi radians.  Angle ranges that include 0rad or 2pi rad do not work.  A rotation must be made to accommodate for this.  The GradientData class now ensures the angles range between 0 and 4pi rad and that the difference between the starting and ending angle is at most 2pi.  It now accounts for angle ranges that have 0 and 2pi rad angles by using a new custom gradient transform, GradientCenterRotation, as well as modifying GradientReversal, formerly known as GradientRotationReverse.