Fix errors in step-by-step letters in font and error in u vowel
Some letters were drawn poorly and led to small lines popping out of the sides of the letter in PracticeCharacterPage.  This is because the current and previous step-by-step letters are stacked on top of eachother.  Some previous letters were drawn so that they would stick out on the final composite letter, and were changed accordingly.  The u vowel was updated so that its sixth anchor point was below the lowest part of the letter.  Before, it would show up as a blank character in DrawnStroke.  Curiously, it was rendered normally in the title of PracticeCharacterPage.