Refactor DrawnStroke for multistep animations, edit GradientData.multistep
DrawnStroke now takes in a step-by-step character to draw or a list of sub stroke characters that make up one step-by-step character.  Each sub stroke is drawn much like a regular, singular step-by-step character.  But after a sub stroke is drawn, it moves onto the other sub stroke, and draws the previously animated one into the background.  The _animation widget has a modified setState that increments characterStepIdx, an index for the current sub stroke in the given sub stroke list.  The animation to draw the whole step-by-step character still takes the same amount of time, but now the animation is segmented into equal parts for each substroke.  The GradientData.multiStep class in step_by_step_stroke_data is modified to send a shadercallback that works with these new changes in DrawnStroke.