Finish Practice animations, Add tsa animation
The error in the previous commit was because the cha and tsa characters are of different character widths because of the final flag-like stroke on the tsa.  When I was making the sub stroke characters for cha, I accidentally used the tsa character as a base, thus causing the error.