Enabled text to be retained in WritingMode when switching between Writing and Practice
Move ChangeNotifierProvider up widget tree so that the text written in writing mode is preserved if the user swaps to practice mode.  Previously, the Navigator built a new WritingMode if the user pressed Writing in the Start Screen.  This would rebuild the ChangeNotifierProvider that was the first widget in WritingMode.  Now, since ChangeNotifierProvider is further up the widget tree, it is nolonger rebuilt everytime, preserving the previous text that was written.